refactor: split db/src/books.rs into focused sub-modules
The post-issue-#83/#202 `books.rs` had grown to 2,537 lines — the single
largest file in the workspace — and combined four distinct responsibilities
(projection helpers, library list/count, single-book lookups, FTS search)
that made it hard to navigate and review.

Apply the same `mod foo;` + sub-directory pattern used in `db/src/auth/`
and `server/src/backend/` to carve `books.rs` into focused topics:

* `books/projection.rs` — `BOOK_COLUMNS`, JSON row decoders, description
  sanitization, the `row_to_ebook` mapper, and `backfill_creator_ids`.
* `books/list.rs` — `list_books`, `list_indexed_rows` + `IndexedRow`,
  `count_books`, and the two `library_from_db*` helpers.
* `books/get.rs` — `get_book`, `get_book_by_uuid`, and
  `resolve_book_id_by_uuid`.
* `books/search.rs` — `search_books`, `search_books_with_total`, and
  `count_search_books`.
* `books/tests.rs` — the existing `#[cfg(test)]` module, kept intact so
  multi-function tests (`list_and_search_books_return_multi_valued_fields`
  and the override-merge cases that touch `get_book` + `list_books`) stay
  co-located.

`books.rs` becomes a 41-line module root that re-exports every previously
public and `pub(crate)` symbol so callers in `server/`, `frontend/`, and
sibling `db/` modules (`discovery`, `palette`, `metadata_overrides`) are
unchanged.

Pure code motion: no behavior changes, no SQL changes, no signature
changes. All 45 `books` tests and the workspace-wide test suite pass.

`sync.rs` (1,730), `discovery.rs` (1,443), and `palette.rs` (1,400)
remain large and are out of scope for this PR — the issue calls them
out as follow-up candidates.

Closes #273
